Refactor: Inputs receives a empty guess struct
